# contracts/openapi/api.v1.yaml
...
components:
  schemas:
...
    RegisterUserRequest:
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: string
      required:
        - username
        - email
        - password
      type: object
    UserRegistrationResponse:
      properties:
        createdAt:
          format: date-time
          type: string
        email:
          type: string
        userId:
          format: uuid
          type: string
        username:
          type: string
        wallet:
          $ref: "#/components/schemas/WalletSummaryDto"
      required:
        - userId
        - username
        - email
        - createdAt
        - wallet
      type: object
...
    WalletSummaryDto:
      properties:
        availableBalance:
          format: double
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
        currency:
          type: string
        reservedBalance:
          format: double
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
        totalBalance:
          format: double
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
      required:
        - currency
        - totalBalance
        - reservedBalance
        - availableBalance
      type: object
info:
  title: TradingSimulator.Api | v1
  version: 1.0.0
openapi: 3.1.1
paths:
...
  /api/users:
    post:
      operationId: RegisterUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRegistrationResponse"
          description: Created
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "422":
          description: Unprocessable Entity
      tags:
        - Users
...